Split up builder.py
Split nengo.builder into a submodule in order to organize
all of the stuff that was accumulating in there.
There may be a slight increase in the number of lines of code,
but I think this makes reading the code a lot easier, and future
edits to the builder will be a lot easier to merge/rebase.

Note that none of the actual logic changed; this was all code moves,
and import changes in order to get everything working.
