+++ nengo/builder/__init__.py
@@ -1,0 +1,9 @@
+from .builder import Builder, Model
+from .connection import build_connection
+from .ensemble import build_ensemble
+from .learning_rules import build_bcm, build_oja, build_pes
+from .network import build_network
+from .neurons import build_lif, build_lifrate, build_alif, build_alifrate
+from .node import build_node
+from .probe import build_probe
+from .synapses import build_filter, build_lowpass, build_alpha
